% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Batch.R
\name{ACHR_batch_V1}
\alias{ACHR_batch_V1}
\title{Batch processing for larger numbers of threads}
\usage{
ACHR_batch_V1(inFileName)
}
\arguments{
\item{inFileName}{name of file (CSV format) containing the raw thread data.}
}
\value{
nothing
}
\description{
ACHR stands for Antecedents of Complexity in Healthcare Routines.  This is function is set up to compute process parameters on thousands of patient visits.
}
