% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{CF_multi_pie_event}
\alias{CF_multi_pie_event}
\title{CF_multi_pie_event}
\usage{
CF_multi_pie_event(o, e, CF, r, zm)
}
\arguments{
\item{o}{data frame of raw occurrences (for the names)}

\item{e}{data frame with events}

\item{CF}{list of contextual factors (columns) to include in the display}

\item{r}{row number of cluster (the number of the event node)}

\item{zm}{integer for zoom column}
}
\value{
plotyly pie charts (one or more)
}
\description{
Make multi-pie for click event from force network layout.  Generate a small plot of context factor pie charts when you click on a node in the graph
}
