% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{cfnames}
\alias{cfnames}
\title{cfnames provides names of all the contextual factors (except the time stamp)}
\usage{
cfnames(o)
}
\arguments{
\item{o}{data frame with threads}
}
\value{
list of column names
}
\description{
cfnames provides names of all the contextual factors (except the time stamp)
}
