% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Mappings.R
\name{check_POV_name}
\alias{check_POV_name}
\title{check_POV_name}
\usage{
check_POV_name(mapname)
}
\arguments{
\item{mapname}{name of map attempting to be created}
}
\value{
TRUE or FALSE
}
\description{
Checks the name attempting to be create against the list of
map names in memory and forces the creation of a new name.
}
