% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{circleVisNetwork}
\alias{circleVisNetwork}
\title{circleVisNetwork}
\usage{
circleVisNetwork(n, directed = "directed", showTitle = FALSE)
}
\arguments{
\item{n}{list with nodeDF and edgeDF dataframes}

\item{directed}{type of network = directed or not}

\item{showTitle}{- show the title or not}
}
\value{
visnetwork object
}
\description{
Produces a circle network layout using visNetwork
}
