% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{cleanOcc}
\alias{cleanOcc}
\title{Title}
\usage{
cleanOcc(o, cfnames)
}
\arguments{
\item{cfnames}{}
}
\description{
Title
}
\seealso{
Other ThreadNet_Misc: \code{\link{cfnames}},
  \code{\link{combineContextFactors}},
  \code{\link{convert_TN_to_TramineR}},
  \code{\link{get_CF_levels}},
  \code{\link{get_moving_window}},
  \code{\link{get_threadList}}, \code{\link{numThreads}},
  \code{\link{read_occurrences}},
  \code{\link{threadSizeTable}}
}
