% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{clusterEvents}
\alias{clusterEvents}
\title{Clusters occurrences or eents}
\usage{
clusterEvents(e, NewMapName, cluster_method, thread_CF, event_CF,
  what_to_return = "POV")
}
\arguments{
\item{e}{a dataframe of events or occurrences}

\item{NewMapName}{= used to store this mapping for visualization and comparison}

\item{cluster_method}{= method for clustering}

\item{thread_CF}{- context factors used to delineate threads}

\item{event_CF}{- context factors used to define events}

\item{what_to_return}{POV or Cluster solution}
}
\value{
event data frame with occurrences aggregated into events or cluster solution
}
\description{
cluster_method is either "Sequential similarity" or "Contextual Similarity" or "Network Structure"
}
