% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{combineContextFactors}
\alias{combineContextFactors}
\title{combineContextFactors creates a new column that combines some set of other columns}
\usage{
combineContextFactors(o, CF, newCol)
}
\arguments{
\item{o}{data frame with threads}

\item{CF}{contextual factors to be combined.}

\item{newCol}{name of the new combined conextual factor}
}
\value{
data frame with the new column
}
\description{
this function adds a new column to the occurrenes table based on a combination of context factors CF). For example, actor+action
}
