% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Metrics.R
\name{compression_index}
\alias{compression_index}
\title{Computes the compressibility of the data in one column of a data frame}
\usage{
compression_index(df, CF)
}
\arguments{
\item{df}{a data frame containing occurrences or events}

\item{CF}{a column or contextual factor in that data frame}
}
\value{
number containing compressibility index, 0 < i < 1
}
\description{
Compressibility is an index of complexity -- more compressible means less complex.  This function computes the ratio of compressed data
to the original data.  Should be between zero and one.  Uses built-in functions for in=memory compression
}
