% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Metrics.R
\name{compute_entropy}
\alias{compute_entropy}
\title{Compute the entropy of a contextual factor}
\usage{
compute_entropy(freq)
}
\arguments{
\item{freq}{is the frequency distribution of the levels in the factor}
}
\value{
number
}
\description{
Each column in the raw data represents a contextual factor.  This function computes the entropy of each factor that is selected for use in the
analysis.
}
\seealso{
Other ThreadNet_Metrics: \code{\link{compression_index}},
  \code{\link{estimate_network_complexity}},
  \code{\link{estimate_task_complexity_index}},
  \code{\link{routineness_metric}}
}
