% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Mappings.R
\name{delete_POV}
\alias{delete_POV}
\title{delete_POV}
\usage{
delete_POV(mapname)
}
\arguments{
\item{mapname}{name of POV map}
}
\value{
None, updates global variables
}
\description{
Deletes all the data assocated with this POV
}
