% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{diff_handoffs}
\alias{diff_handoffs}
\title{diff_handoffs}
\usage{
diff_handoffs(o)
}
\arguments{
\item{o}{dataframe of occurrences}
}
\value{
o dataframe of occurrences with handoff count filled in
}
\description{
count the handoffs, but reverse coded -- zero = all different
}
