% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{diff_tStamp}
\alias{diff_tStamp}
\title{diff_tStamp}
\usage{
diff_tStamp(ts)
}
\arguments{
\item{ts}{column of time stamps}
}
\value{
Column of differences between timestamps
}
\description{
Time gaps -- just pass in the column of time stamps - uses auto units
}
