% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{dual_window_correlation}
\alias{dual_window_correlation}
\title{dual_window_correlation}
\usage{
dual_window_correlation(e, w, s = 1, n = 2)
}
\arguments{
\item{e}{data fraom for POV}

\item{w}{width of moving window}

\item{s}{step - how far to move window in each increment (default is 1)}

\item{n}{number of windows (default is 2)}
}
\description{
similar to moving window, except one window on each side of a focal thread.
}
