% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{dualmovingWindowCorrelation}
\alias{dualmovingWindowCorrelation}
\title{dualmovingWindowCorrelation}
\usage{
dualmovingWindowCorrelation(trace)
}
\arguments{
\item{trace}{list of (x,y) coordinates}
}
\value{
plotly scatter plot
}
\description{
Creates plotly diagram showing correlation of moving windows across time
This version computes the correlation of TWO adjacent windows that never overlap
}
