% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Metrics.R
\name{estimate_task_complexity_index}
\alias{estimate_task_complexity_index}
\title{Estimates the number of paths in a directed graph}
\usage{
estimate_task_complexity_index(v, e)
}
\arguments{
\item{v}{number of vertices (or nodes)}

\item{e}{number of edges}
}
\value{
number
}
\description{
Same as estimate_network_complexity, but takes this version takes vertices and edges as parameters
}
