% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{eventNetwork}
\alias{eventNetwork}
\title{Circular network layout for event network (USES visnetwork)}
\usage{
eventNetwork(et, TN, CF, timesplit)
}
\arguments{
\item{et}{dataframe with the threads to be graphed}

\item{TN}{the column with the threadNumber}

\item{CF}{is the contetual factors (column)}

\item{timesplit}{time measure}
}
\value{
plotly object
}
\description{
Should be replaced with a more expressive layout in plotly
}
\seealso{
Other ThreadNet_Graphics: \code{\link{CF_multi_pie}},
  \code{\link{Comparison_Plots}},
  \code{\link{circleVisNetwork}},
  \code{\link{filter_network_edges}},
  \code{\link{forceNetworkD3}}, \code{\link{ng_bar_chart}},
  \code{\link{normalNetwork}}, \code{\link{role_map}},
  \code{\link{threadMap}}, \code{\link{threadTrajectory}}
}
