% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Mappings.R
\name{export_POV}
\alias{export_POV}
\title{export_POV}
\usage{
export_POV(mapname)
}
\arguments{
\item{mapname}{name of POV map}
}
\value{
(writes Rdata file)
}
\description{
Exports the data assocated with this POV as Rdata
}
