% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Mappings.R
\name{export_POV_csv}
\alias{export_POV_csv}
\title{export_POV_csv}
\usage{
export_POV_csv(mapname)
}
\arguments{
\item{mapname}{name of POV map}
}
\value{
(writes CSV file)
}
\description{
Exports the data assocated with this POV as CSV
}
