% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{forceNetworkD3}
\alias{forceNetworkD3}
\title{forceNetworkD3 is an Interactive layout for event network}
\usage{
forceNetworkD3(n)
}
\arguments{
\item{n}{= list with data frames for nodes and edges}
}
\value{
networkD3 object
}
\description{
This produces a force layout network using networkD3
}
