% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{frequent_ngrams}
\alias{frequent_ngrams}
\title{frequent_ngrams}
\usage{
frequent_ngrams(e, TN, CF, minN, maxN, onlyMaximal = TRUE)
}
\arguments{
\item{e}{event data}

\item{TN}{threadNum}

\item{CF}{context factor (column) to look at}

\item{minN}{miniumum ngram length}

\item{maxN}{maximum ngram length}

\item{onlyMaximal}{Filters out ngrams that are included in longer ngrams.  Default is true.}
}
\value{
dataframe of ngrams
}
\description{
combined set of frequent ngrams within a range o lengths
}
