% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Mappings.R
\name{get_POV_COMPARISON_CF}
\alias{get_POV_COMPARISON_CF}
\title{get_POV_COMPARISON_CF}
\usage{
get_POV_COMPARISON_CF(mapname, CF_list)
}
\arguments{
\item{mapname}{name of POV map}

\item{CF_list}{list of other column names}
}
\value{
comparison CFs for that POV
}
\description{
Gets the CFs that can be used for comparisons in this POV
}
