% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Mappings.R
\name{get_POV_EVENT_CF}
\alias{get_POV_EVENT_CF}
\title{get_POV_EVENT_CF}
\usage{
get_POV_EVENT_CF(mapname)
}
\arguments{
\item{mapname}{name of POV map}
}
\value{
event CFs for that POV
}
\description{
Gets the CFs that define events in this POV
}
