% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Mappings.R
\name{get_POV_name_list}
\alias{get_POV_name_list}
\title{get_POV_name_list}
\usage{
get_POV_name_list()
}
\value{
List of POV names
}
\description{
Get list of POV names for all of the dropdown boxes on the UI
}
