% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{get_threadList}
\alias{get_threadList}
\title{get_threadList}
\usage{
get_threadList(e, TN, SN)
}
\arguments{
\item{e}{data frame with threaded events}

\item{TN}{Column with threadNumber}

\item{SN}{Column with sequence numbers}
}
\value{
list of thread numbers
}
\description{
get_threadList returns a list of all thread numbers
}
