% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{make_subsets}
\alias{make_subsets}
\title{make_subsets}
\usage{
make_subsets(d, n)
}
\arguments{
\item{d}{data frame with occurrences or events}

\item{n}{number of buckets}
}
\value{
list of smaller data frames
}
\description{
this function is used to split up the threads into n ~equal buckets
}
