% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{movingWindowCorrelation}
\alias{movingWindowCorrelation}
\title{movingWindowCorrelation}
\usage{
movingWindowCorrelation(trace)
}
\arguments{
\item{trace}{list of (x,y) coordinates}
}
\value{
plotly scatter plot
}
\description{
Creates plotly diagram showing correlation of moving windows across time
In this version, the moving window can overlap with itself as it slides
}
