% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{newColName}
\alias{newColName}
\title{newColName}
\usage{
newColName(CF_list)
}
\arguments{
\item{CF_list}{list of context factors to define events}
}
\value{
column name
}
\description{
returns new combined column name from a list of CFs
}
