% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{ng_bar_chart}
\alias{ng_bar_chart}
\title{Create an ngram bar chart}
\usage{
ng_bar_chart(o, TN, CF, n, mincount)
}
\arguments{
\item{o}{a dataframe of occurrences or events}

\item{TN}{the column that contains the threadNum}

\item{CF}{the contextual factor within which to count the n-grams}

\item{n}{the length of the ngram}

\item{mincount}{the minimum count to display}
}
\value{
plotly object
}
\description{
Shows the n-grams within a set of threads (but not splitting across threads). This provides a visual indication of how repetitive the threads are.
}
\seealso{
Other ThreadNet_Graphics: \code{\link{CF_multi_pie}},
  \code{\link{Comparison_Plots}},
  \code{\link{eventNetwork}}, \code{\link{forceNetworkD3}},
  \code{\link{threadMap}}
}
