% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{normalNetwork}
\alias{normalNetwork}
\title{normalNetwork}
\usage{
normalNetwork(e, o, cf)
}
\arguments{
\item{e}{event data frame}

\item{o}{occurrence data frame}

\item{cf}{context factor for the graph}
}
\value{
visnetwork object
}
\description{
Produced a network of co-occurrences for any given CF and displays it in a visNetwork circle layout
}
