% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{OccToEvents}
\alias{OccToEvents}
\title{Maps occurrences into events}
\usage{
OccToEvents(o, mapping = "One-to-One", m = "Variable chunks",
  uniform_chunk_size = 1, tThreshold = 1, chunk_CF, EventMapName, EVENT_CF,
  compare_CF, timescale)
}
\arguments{
\item{o}{a dataframe of occurrences}

\item{mapping}{= one-to-one or clustering}

\item{m}{= method parameter = one of c('Variable chunks','Uniform chunks')}

\item{uniform_chunk_size}{= used to identify breakpoints -- from input slider}

\item{tThreshold}{= used to identify breakpoints -- from input slider}

\item{chunk_CF}{- context factors used to delineate chunks}

\item{EventMapName}{= used to store this mapping in an environment (not used yet)}

\item{EVENT_CF}{- context factors used to define events}

\item{compare_CF}{= context factors used for comparison -- need to be copied over here when the thread is created.}

\item{timescale}{hours, min or sec}
}
\description{
Thus function provides a place to map occurrences into events, so is is not necessary to interpret individual
occurrences in isolation.  There are many ways to accomplish this mapping.
}
\seealso{
Other ThreadNet_Core: \code{\link{ThreadOccByPOV}},
  \code{\link{count_ngrams}},
  \code{\link{threads_to_network}}
}
