% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{OccToEvents3}
\alias{OccToEvents3}
\title{OccToEvents3}
\usage{
OccToEvents3(o, EventMapName, THREAD_CF, EVENT_CF, compare_CF, TN, CF, rx,
  KeepIrregularEvents)
}
\arguments{
\item{o}{a dataframe of occurrences}

\item{EventMapName}{= used to store this mapping for visualization and comparison}

\item{THREAD_CF}{- context factors used to delineate threads}

\item{EVENT_CF}{- context factors used to define events}

\item{compare_CF}{= context factors used for comparison -- need to be copied over here when the thread is created.}

\item{TN}{ThreadNum}

\item{CF}{context factor}

\item{rx}{list of patterns}

\item{KeepIrregularEvents}{= keep or drop events that don't fit patterns}
}
\value{
event data frame, with occurrences aggregated into events.
}
\description{
Creates events based on frequent ngrams or regular expressions
}
