% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{OccToEvents_By_Chunk}
\alias{OccToEvents_By_Chunk}
\title{Maps occurrences into events}
\usage{
OccToEvents_By_Chunk(o, m, EventMapName, uniform_chunk_size, tThreshold,
  timescale = "mins", chunk_CF, thread_CF, event_CF, compare_CF)
}
\arguments{
\item{o}{a dataframe of occurrences}

\item{m}{= method parameter = one of c('Variable chunks','Uniform chunks')}

\item{EventMapName}{= used to store this mapping for visualization and comparison}

\item{uniform_chunk_size}{= used to identify breakpoints -- from input slider}

\item{tThreshold}{= used to identify breakpoints -- from input slider}

\item{timescale}{hours, min or sec}

\item{chunk_CF}{- context factors used to delineate chunks}

\item{thread_CF}{- context factors used to delineate threads}

\item{event_CF}{- context factors used to define events}

\item{compare_CF}{= context factors used for comparison -- need to be copied over here when the thread is created.}
}
\value{
event data frame, with occurrences aggregated into events.
}
\description{
Thus function provides a place to map occurrences into events, so is is not necessary to interpret individual
occurrences in isolation.  There are many ways to accomplish this mapping.
}
\seealso{
Other ThreadNet_Core: \code{\link{OccToEvents3}},
  \code{\link{ThreadOccByPOV}},
  \code{\link{clusterEvents}}, \code{\link{count_ngrams}},
  \code{\link{threads_to_network_original}}
}
