% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Metrics.R
\name{plot_entropy}
\alias{plot_entropy}
\title{plot entropy as a function of zoom_level}
\usage{
plot_entropy(e)
}
\arguments{
\item{e}{data from of events with zoom levels}
}
\value{
regular R plot
}
\description{
Gets the zoom levels, grep out the 'Z_' column names...
}
