% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{role_map}
\alias{role_map}
\title{role_map}
\usage{
role_map(e, o, cfs)
}
\arguments{
\item{e}{event data frame}

\item{o}{occurrence data frame}

\item{cfs}{context factors for the graph}
}
\value{
plotly heat map
}
\description{
A role map (like a heat map) that will show "who does what" for any set of events
}
