% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event_Mappings.R
\name{store_POV}
\alias{store_POV}
\title{store_POV}
\usage{
store_POV(EventMapName, e, thread_CF, event_CF)
}
\arguments{
\item{EventMapName}{name of map attempting to be created}

\item{e}{data frame with POV to be stored}

\item{thread_CF}{List of CFs to be stored}

\item{event_CF}{List of CFs to be stored}
}
\value{
None, updates global variables
}
\description{
Stores the POV and context factors
}
