% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{support_level}
\alias{support_level}
\title{support_level}
\usage{
support_level(tv, ng)
}
\arguments{
\item{tv}{text vector of threads}

\item{ng}{ngram to be located in the threads}
}
\value{
percentage of threads containing the ngram
}
\description{
Counts what fraction of the threads a particular ngram appears in
}
