% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{thread_text_vector}
\alias{thread_text_vector}
\title{thread_text_vector}
\usage{
thread_text_vector(o, TN, CF, delimiter)
}
\arguments{
\item{o}{a dataframe of events or occurrences}

\item{TN}{= threadNum}

\item{CF}{= CF or columm to include}

\item{delimiter}{usually comma or blank}
}
\value{
vector of threads as delimited character strings
}
\description{
Create a vector of threads
}
