% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{threadMap}
\alias{threadMap}
\title{Shows threads in a horizontal layout}
\usage{
threadMap(or, TN, timescale, CF, shape)
}
\arguments{
\item{or}{Dataframe of threads}

\item{TN}{name of column with thread number}

\item{timescale}{name of column that will be used to plot x-axis of events. It can be the can be the time stamp (for clock time) or the sequence number (for event time)}

\item{CF}{name of contextual factor that will determine the colors}

\item{shape}{shape code for the markers on the threadmap}
}
\value{
plotly object
}
\description{
Creates a plotly chart of threads in either clock time or event time, depending on the timescale parameter.
}
\seealso{
Other ThreadNet_Graphics: \code{\link{CF_multi_pie}},
  \code{\link{Comparison_Plots}},
  \code{\link{circleVisNetwork}},
  \code{\link{eventNetwork}},
  \code{\link{filter_network_edges}},
  \code{\link{forceNetworkD3}}, \code{\link{ng_bar_chart}},
  \code{\link{normalNetwork}}, \code{\link{role_map}},
  \code{\link{threadTrajectory}}
}
