% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{ThreadOccByPOV}
\alias{ThreadOccByPOV}
\title{Make new threads from a new POV}
\usage{
ThreadOccByPOV(o, THREAD_CF, EVENT_CF)
}
\arguments{
\item{o}{is the dataframe of cleaned ocurrences}

\item{THREAD_CF}{is a list of 1 or more context factors that define the threads (and stay constant during each thread)}

\item{EVENT_CF}{is a list of 1 or more context factors that define events (and change during threads)}
}
\value{
dataframe containing the same occurrences sorted from a different point of view
}
\description{
Take the raw occurrences from the input file and sort them by time stamp within
a set of contextual factors that remain constant for each thread.
}
