% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Core.R
\name{threads_to_network}
\alias{threads_to_network}
\title{Converts threads to network}
\usage{
threads_to_network(et, TN, CF, timesplit)
}
\arguments{
\item{et}{dataframe containing threads}

\item{TN}{name of column in dataframe that contains a unique thread number for each thread}

\item{CF}{name of the column in dataframe that contains the events that will form the nodes of the network}

\item{timesplit}{time measure}
}
\value{
a list containing two dataframes, one for the nodes (nodeDF) and one for the edges (edgeDF)
}
\description{
Converts a sequentially ordered streams of ;events (threads) and creates a unimodal, unidimensional network.
Sequentially adjacent pairs of events become edges in the resulting network.
}
\seealso{
Other ThreadNet_Core: \code{\link{OccToEvents_By_Chunk}},
  \code{\link{ThreadOccByPOV}}, \code{\link{count_ngrams}}
}
