% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{threadTrajectory}
\alias{threadTrajectory}
\title{threadTrajectory}
\usage{
threadTrajectory(or)
}
\arguments{
\item{or}{event data frame}
}
\value{
plotly scatter plot
}
\description{
Create a plotly diagram showing relative time versus sequential time.
Inspired by Gergen and Danner-Schroeder
}
