% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{threshold_slider_max}
\alias{threshold_slider_max}
\title{threshold_slider_max}
\usage{
threshold_slider_max(o)
}
\arguments{
\item{o}{dataframe of occurrences}
}
\value{
slider max
}
\description{
These were used on the occ-to-event tab to configure the slider
}
