% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{window_correlation}
\alias{window_correlation}
\title{window_correlation}
\usage{
window_correlation(e, w, s = 1, n = 2)
}
\arguments{
\item{e}{data fraom for POV}

\item{w}{width of moving window}

\item{s}{step - how far to move window in each increment (default is 1)}

\item{n}{number of windows (default is 2)}
}
\description{
Correlation between moving windows
}
