% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{zoom_upper_limit}
\alias{zoom_upper_limit}
\title{zoom_upper_limit}
\usage{
zoom_upper_limit(e)
}
\arguments{
\item{e}{data frame}
}
\value{
biggest zoom level
}
\description{
find the biggest column with ZM_, and then get the number that goes with that.

Used to set upper limit on sliders for zooming
}
