% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{zoomColumn}
\alias{zoomColumn}
\title{zoomColumn}
\usage{
zoomColumn(z)
}
\arguments{
\item{z}{integer for the zoom level}
}
\value{
column name for that zoom level ('ZM_n)
}
\description{
This function takes a slider value and returns a valid column name for zooming
if the argument is null, then use ZM_1
}
