% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{Comparison_Plots}
\alias{Comparison_Plots}
\title{Title}
\usage{
Comparison_Plots(e, CF, CF_levels, nTimePeriods = 1, ng_size, zoom_level)
}
\arguments{
\item{zoom_level}{}
}
\description{
Title
}
