% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{Comparison_Plots}
\alias{Comparison_Plots}
\title{Comparison plots}
\usage{
Comparison_Plots(e, o, CF, CF_levels, nTimePeriods = 1, plot_type,
  role_map_cfs)
}
\arguments{
\item{e}{dataframe with threads to be plotted}

\item{o}{dataframe with the original data}

\item{CF}{contextul factors}

\item{CF_levels}{list of levels from whicheve contextual factor was chosen for comprisons (e.g., location =1, 2, 3)}

\item{nTimePeriods}{how many time periods to divide the data?}

\item{plot_type}{a type of plotly plot with a function written}

\item{role_map_cfs}{context factors for the role map plot}
}
\value{
plotly object, including subplots
}
\description{
Produce a set set of comparison sub-plots in an array.  Ideally, we should be able to use any of the plots. So far it is only bar charts.
This is a prototype that could use rather extensive redesign...
}
\seealso{
Other ThreadNet_Graphics: \code{\link{CF_multi_pie}},
  \code{\link{circleVisNetwork}},
  \code{\link{eventNetwork}},
  \code{\link{filter_network_edges}},
  \code{\link{forceNetworkD3}}, \code{\link{ng_bar_chart}},
  \code{\link{normalNetwork}}, \code{\link{role_map}},
  \code{\link{threadMap}}, \code{\link{threadTrajectory}}
}
