% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Graphics.R
\name{forceNetworkD3}
\alias{forceNetworkD3}
\title{NetworkD3 layout for event network}
\usage{
forceNetworkD3(n)
}
\arguments{
\item{n}{= list with data frames for nodes and edges}
}
\value{
networkD3 object
}
\description{
NetworkD3 layout for event network
}
\seealso{
Other ThreadNet_Graphics: \code{\link{CF_multi_pie}},
  \code{\link{Comparison_Plots}},
  \code{\link{eventNetwork}}, \code{\link{ng_bar_chart}},
  \code{\link{threadMap}}
}
