% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreadNet_Misc.R
\name{threadSizeTable}
\alias{threadSizeTable}
\title{threadSizeTable provides a distribution of the length of threads}
\usage{
threadSizeTable(o, TN)
}
\arguments{
\item{o}{data frame with threads}

\item{TN}{column comtaining the threadNumber}
}
\value{
data frame with table of thread lengths
}
\description{
This function should work on either ocurrences or events.
it returns length and duration of each thread.It requires tStamp field to compute duration.
}
